* commit 'aa060b64ea12a90a3fd0f30971101a42aa90cc27':
  Changed semantic working.

+++ src/mips/RegisterManager.java
@@ -84,7 +84,7 @@ public class RegisterManager {
             if (matcher.matches()) {
                 int label = Integer.parseInt(matcher.group(1));
                 if (activeLoops.contains(label)) {
-                    activeLoops.remove(label);
+                    activeLoops.remove((Integer) label);
 
                     // Add new line
                     ArrayList<String> registers = new ArrayList<>();
